use modified version of Cassandra image to fix volume mounts
I have been trying to use an emptyDir volume mount for Cassandra's conf
directory so that it is visible to Medusa. The upstream image
https://github.com/docker-library/cassandra/blob/master/3.11/Dockerfile moves
files to /etc/cassandra. The problem is that after that is done, the volume
gets mounted, and the files are no longer visible. I moved the code for setting
up /etc/cassandra into docker-entrypoint.sh which runs after volumes are
mounted. Problem solved!