initial commit for cassandra-init image
I am currently doing a PoC for an in-place restore with Medusa. The restore
operation runs in an init container. Medusa needs access to Cassandra's config
files, namely cassandra.yaml, so I had to pull out the bits that set up the
configuration in the Cassandra image and put it into a separate image that runs
in an init container prior to the restore container.